KatAppOptions.Framework, injectTemplatesWithoutSource, Documentation
KatAppOptions.Framework - allows client hosting KatApp and calling RBLe to specify the framework being used.  If no framework is provided, it is assumed that it is legacy RBLe.ts.

injectTemplatesWithoutSource - fixed code that tries to filter out the container if it is a template (so that it wasn't called twice).

Documentation - Documented /unique-summary and table nesting []