NavigationInputs, rbl-config CalcEngine, NextCalculation settings, createModalDialog
Changed NavigationInputs to be placed in ManualInputs so that they are applied for every calculation while on page.  Before they were on DefaultInputs so were applied for first calculation, then discarded.

rbl-config/calcengine will use 'default' as the key if no name is provided.

NextCalculation - saveLocations, expireCache, and trace are used instead of data attributes.  Can add multiple saveLocations and also specify server only or not if desired

createModalDialog - Renamed createConfirmDialog so that UIUtilities exposes this method to display a modal whenever KatApp events want to.